Use proper ANSI code "39" for default color
The method of querying "isDarkTerminal()" via the environment variable
COLORFGBG is broken and not supported by all terminals.

The ANSI color code "39" denotes the default foreground color in any
properly configured terminal, and should _always_ be readable
on the default background, both for light- and dark-background terminals.

It has the same effect as code "0" (normal/reset), but without
resetting other text attributes (like bold or underlined).